Add Storage 클래스
Storage 클래스 추가 후, 해당 클래스에 대응하도록 코드를 수정함.
