+++ main.py
@@ -1,4 +1,5 @@
 from Player import Player, Death
+from Storage import Storage
 
 from initializeMaintenance import initializeMaintenance # 기초 정비
 from explore import explore # 탐사 출발
@@ -6,12 +7,14 @@ from event import event # 이벤트
 from comeback import comeback # 쉘터 복귀
 from maintenance import maintenance # 정비
 
+
 def main():
     """프로그램 시작"""
     print("프로그램 시작")
 
     player = Player()
     days = int(0) # 지난 날짜 수
+    shelterInventory = Storage()
 
     try:
         initializeMaintenance(player)
@@ -24,7 +27,7 @@ def main():
 
                     if comeback(player): # 쉘터에 복귀하면
                         break
-                maintenance(player) # 정비할 수 있음
+                maintenance(player, shelterInventory) # 정비할 수 있음
     except Death as e:
         print(e)
     except:
